Move profile stuff into profile folder and namespace.
